Feature/db pools

See merge request ACVTS/GenVals/gen-val!128

+++ src/pools/src/NIST.CVP.Pools/Models/PoolInformation.cs
@@ -3,6 +3,6 @@
     public class PoolInformation
     {
         public bool PoolExists { get; set; } = true;
-        public int FillLevel { get; set; }
+        public long FillLevel { get; set; }
     }
 }
